fix: move create invoice to module
